# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Athletes Survey API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
}

export interface Demographics {
  fullName: string;
  age: number;
  competitionAgeGroup: string;
  gender: string;
  height: number;
  weight: number;
  city: string;
  maritalStatus: string;
  sport: string;
  nationalTeamHistory: string;
  premierLeagueHistory: string;
  firstLeagueHistory: string;
  yearsOfExperience: number;
  injuryHistory: string;
  /** @nullable */
  timeSinceLastInjury?: string | null;
  physiotherapyTreatment: string;
  surgerHistory: string;
  /** @nullable */
  injuryDetails?: string | null;
}

export interface CreateSubmissionBody {
  demographics: Demographics;
  /** Aggression questionnaire answers (12 questions, scale 1-5) */
  questionnaire1: number[];
  /** Burnout questionnaire answers (11 questions, scale 1-5) */
  questionnaire2: number[];
  /** Exercise addiction questionnaire answers (6 questions, scale 1-5) */
  questionnaire3: number[];
  /** Competitive anxiety questionnaire answers (15 questions, scale 1-3) */
  questionnaire4: number[];
  /** Mental toughness questionnaire answers (14 questions, scale 1-5) */
  questionnaire5: number[];
  /** Mood states questionnaire answers (32 questions, scale 1-5) */
  questionnaire6: number[];
  /** Sport participation motivation questionnaire answers (30 questions, scale 1-3) */
  questionnaire7: number[];
}

export interface SubmissionSummary {
  id: number;
  fullName: string;
  age: number;
  sport: string;
  city: string;
  gender: string;
  submittedAt: string;
}

export interface Submission {
  id: number;
  demographics: Demographics;
  questionnaire1: number[];
  questionnaire2: number[];
  questionnaire3: number[];
  questionnaire4: number[];
  questionnaire5: number[];
  questionnaire6: number[];
  questionnaire7: number[];
  submittedAt: string;
}

export interface ListSubmissionsResponse {
  submissions: SubmissionSummary[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface DemographicBreakdown {
  label: string;
  count: number;
}

export interface AdminStats {
  totalSubmissions: number;
  totalThisWeek: number;
  totalThisMonth: number;
  sportBreakdown: DemographicBreakdown[];
  cityBreakdown: DemographicBreakdown[];
  ageGroupBreakdown: DemographicBreakdown[];
  maritalStatusBreakdown: DemographicBreakdown[];
  avgAge: number;
  avgYearsExperience: number;
  injuryRate: number;
  nationalTeamRate: number;
}

export type ListAdminSubmissionsParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  sport?: string | null;
  /**
   * @nullable
   */
  city?: string | null;
};