# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Athletes Survey API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: submissions
    description: Survey submission operations
  - name: admin
    description: Admin panel operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit survey answers
      description: Save the complete survey submission including demographics and all 7 questionnaires
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Authenticate as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get aggregate statistics
      description: Returns total participants, completion rates, and demographic aggregates
      security:
        - adminSession: []
      responses:
        "200":
          description: Stats returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions:
    get:
      operationId: listAdminSubmissions
      tags: [admin]
      summary: List all submissions
      description: Returns paginated list of all survey submissions with basic demographics
      security:
        - adminSession: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sport
          in: query
          schema:
            type: ["string", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSubmissionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions/{id}:
    get:
      operationId: getAdminSubmission
      tags: [admin]
      summary: Get individual submission
      description: Returns full details of a specific submission including all questionnaire answers
      security:
        - adminSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/export:
    get:
      operationId: exportSubmissions
      tags: [admin]
      summary: Export all submissions as CSV
      description: Returns all survey data as a downloadable CSV file suitable for SPSS
      security:
        - adminSession: []
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Demographics:
      type: object
      properties:
        fullName:
          type: string
        age:
          type: integer
        competitionAgeGroup:
          type: string
        gender:
          type: string
        height:
          type: number
        weight:
          type: number
        city:
          type: string
        maritalStatus:
          type: string
        sport:
          type: string
        nationalTeamHistory:
          type: string
        premierLeagueHistory:
          type: string
        firstLeagueHistory:
          type: string
        yearsOfExperience:
          type: integer
        injuryHistory:
          type: string
        timeSinceLastInjury:
          type: ["string", "null"]
        physiotherapyTreatment:
          type: string
        surgerHistory:
          type: string
        injuryDetails:
          type: ["string", "null"]
      required:
        - fullName
        - age
        - competitionAgeGroup
        - gender
        - height
        - weight
        - city
        - maritalStatus
        - sport
        - nationalTeamHistory
        - premierLeagueHistory
        - firstLeagueHistory
        - yearsOfExperience
        - injuryHistory
        - physiotherapyTreatment
        - surgerHistory

    CreateSubmissionBody:
      type: object
      properties:
        demographics:
          $ref: "#/components/schemas/Demographics"
        questionnaire1:
          type: array
          items:
            type: integer
          description: "Aggression questionnaire answers (12 questions, scale 1-5)"
        questionnaire2:
          type: array
          items:
            type: integer
          description: "Burnout questionnaire answers (11 questions, scale 1-5)"
        questionnaire3:
          type: array
          items:
            type: integer
          description: "Exercise addiction questionnaire answers (6 questions, scale 1-5)"
        questionnaire4:
          type: array
          items:
            type: integer
          description: "Competitive anxiety questionnaire answers (15 questions, scale 1-3)"
        questionnaire5:
          type: array
          items:
            type: integer
          description: "Mental toughness questionnaire answers (14 questions, scale 1-5)"
        questionnaire6:
          type: array
          items:
            type: integer
          description: "Mood states questionnaire answers (32 questions, scale 1-5)"
        questionnaire7:
          type: array
          items:
            type: integer
          description: "Sport participation motivation questionnaire answers (30 questions, scale 1-3)"
      required:
        - demographics
        - questionnaire1
        - questionnaire2
        - questionnaire3
        - questionnaire4
        - questionnaire5
        - questionnaire6
        - questionnaire7

    SubmissionSummary:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        age:
          type: integer
        sport:
          type: string
        city:
          type: string
        gender:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - age
        - sport
        - city
        - gender
        - submittedAt

    Submission:
      type: object
      properties:
        id:
          type: integer
        demographics:
          $ref: "#/components/schemas/Demographics"
        questionnaire1:
          type: array
          items:
            type: integer
        questionnaire2:
          type: array
          items:
            type: integer
        questionnaire3:
          type: array
          items:
            type: integer
        questionnaire4:
          type: array
          items:
            type: integer
        questionnaire5:
          type: array
          items:
            type: integer
        questionnaire6:
          type: array
          items:
            type: integer
        questionnaire7:
          type: array
          items:
            type: integer
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - demographics
        - questionnaire1
        - questionnaire2
        - questionnaire3
        - questionnaire4
        - questionnaire5
        - questionnaire6
        - questionnaire7
        - submittedAt

    ListSubmissionsResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - submissions
        - total
        - page
        - limit
        - totalPages

    DemographicBreakdown:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
      required:
        - label
        - count

    AdminStats:
      type: object
      properties:
        totalSubmissions:
          type: integer
        totalThisWeek:
          type: integer
        totalThisMonth:
          type: integer
        sportBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DemographicBreakdown"
        cityBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DemographicBreakdown"
        ageGroupBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DemographicBreakdown"
        maritalStatusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DemographicBreakdown"
        avgAge:
          type: number
        avgYearsExperience:
          type: number
        injuryRate:
          type: number
        nationalTeamRate:
          type: number
      required:
        - totalSubmissions
        - totalThisWeek
        - totalThisMonth
        - sportBreakdown
        - cityBreakdown
        - ageGroupBreakdown
        - maritalStatusBreakdown
        - avgAge
        - avgYearsExperience
        - injuryRate
        - nationalTeamRate