# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Athletes Survey API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Save the complete survey submission including demographics and all 7 questionnaires
 * @summary Submit survey answers
 */
export const CreateSubmissionBody = zod.object({
  demographics: zod.object({
    fullName: zod.string(),
    age: zod.number(),
    competitionAgeGroup: zod.string(),
    gender: zod.string(),
    height: zod.number(),
    weight: zod.number(),
    city: zod.string(),
    maritalStatus: zod.string(),
    sport: zod.string(),
    nationalTeamHistory: zod.string(),
    premierLeagueHistory: zod.string(),
    firstLeagueHistory: zod.string(),
    yearsOfExperience: zod.number(),
    injuryHistory: zod.string(),
    timeSinceLastInjury: zod.string().nullish(),
    physiotherapyTreatment: zod.string(),
    surgerHistory: zod.string(),
    injuryDetails: zod.string().nullish(),
  }),
  questionnaire1: zod
    .array(zod.number())
    .describe("Aggression questionnaire answers (12 questions, scale 1-5)"),
  questionnaire2: zod
    .array(zod.number())
    .describe("Burnout questionnaire answers (11 questions, scale 1-5)"),
  questionnaire3: zod
    .array(zod.number())
    .describe(
      "Exercise addiction questionnaire answers (6 questions, scale 1-5)",
    ),
  questionnaire4: zod
    .array(zod.number())
    .describe(
      "Competitive anxiety questionnaire answers (15 questions, scale 1-3)",
    ),
  questionnaire5: zod
    .array(zod.number())
    .describe(
      "Mental toughness questionnaire answers (14 questions, scale 1-5)",
    ),
  questionnaire6: zod
    .array(zod.number())
    .describe("Mood states questionnaire answers (32 questions, scale 1-5)"),
  questionnaire7: zod
    .array(zod.number())
    .describe(
      "Sport participation motivation questionnaire answers (30 questions, scale 1-3)",
    ),
});

/**
 * Authenticate as admin
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
});

/**
 * Returns total participants, completion rates, and demographic aggregates
 * @summary Get aggregate statistics
 */
export const GetAdminStatsResponse = zod.object({
  totalSubmissions: zod.number(),
  totalThisWeek: zod.number(),
  totalThisMonth: zod.number(),
  sportBreakdown: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
  cityBreakdown: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
  ageGroupBreakdown: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
  maritalStatusBreakdown: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
  avgAge: zod.number(),
  avgYearsExperience: zod.number(),
  injuryRate: zod.number(),
  nationalTeamRate: zod.number(),
});

/**
 * Returns paginated list of all survey submissions with basic demographics
 * @summary List all submissions
 */
export const listAdminSubmissionsQueryPageDefault = 1;
export const listAdminSubmissionsQueryLimitDefault = 20;

export const ListAdminSubmissionsQueryParams = zod.object({
  page: zod.coerce.number().default(listAdminSubmissionsQueryPageDefault),
  limit: zod.coerce.number().default(listAdminSubmissionsQueryLimitDefault),
  sport: zod.coerce.string().nullish(),
  city: zod.coerce.string().nullish(),
});

export const ListAdminSubmissionsResponse = zod.object({
  submissions: zod.array(
    zod.object({
      id: zod.number(),
      fullName: zod.string(),
      age: zod.number(),
      sport: zod.string(),
      city: zod.string(),
      gender: zod.string(),
      submittedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * Returns full details of a specific submission including all questionnaire answers
 * @summary Get individual submission
 */
export const GetAdminSubmissionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAdminSubmissionResponse = zod.object({
  id: zod.number(),
  demographics: zod.object({
    fullName: zod.string(),
    age: zod.number(),
    competitionAgeGroup: zod.string(),
    gender: zod.string(),
    height: zod.number(),
    weight: zod.number(),
    city: zod.string(),
    maritalStatus: zod.string(),
    sport: zod.string(),
    nationalTeamHistory: zod.string(),
    premierLeagueHistory: zod.string(),
    firstLeagueHistory: zod.string(),
    yearsOfExperience: zod.number(),
    injuryHistory: zod.string(),
    timeSinceLastInjury: zod.string().nullish(),
    physiotherapyTreatment: zod.string(),
    surgerHistory: zod.string(),
    injuryDetails: zod.string().nullish(),
  }),
  questionnaire1: zod.array(zod.number()),
  questionnaire2: zod.array(zod.number()),
  questionnaire3: zod.array(zod.number()),
  questionnaire4: zod.array(zod.number()),
  questionnaire5: zod.array(zod.number()),
  questionnaire6: zod.array(zod.number()),
  questionnaire7: zod.array(zod.number()),
  submittedAt: zod.coerce.date(),
});